feat : UserService (추후 Spring Security를 적용시켜 로그인 기능 구현)

+++ src/main/java/com/example/board/service/UserService.java
@@ -26,7 +26,7 @@ public class UserService {
         return userRepository.findAll();
     }
 
-    public User findUser(int id) {
+    public User findUser(Long id) {
         return userRepository.findById(id).orElseThrow(UserNotFoundException::new);
     }
 }
